// npm/packages/utils/generated-types/MinHelseEntities.d.ts
export type ISO8601 = string;
// ...
export interface Samtykke {
  Type: SamtykkeType;
  PersonvernInnstillingDefinisjonGuid: any;
  Opprettet: ISO8601;
  StatusType: SamtykkeStatusType;
  SamtykketAvProfilGuid: any;
  SamtykketAvFornavn: string;
  SamtykketAvEtternavn: string;
  RepresentasjonforholdType: RepresentasjonforholdType;
  StatusKodeType: StatusKodeType;
  ErAktiv: boolean;
}